Mas bugs

+++ Controllers/ReserveController.php
@@ -66,8 +66,13 @@ class ReserveController
     public function RefuseReserve($reserveId)
     {
         $reserve = $this->reserveDAO->GetReserve($reserveId);
-        $reserve->setKeeperReviewStatus(REJECTED);
-        $this->reserveDAO->Update($reserve);
+
+        if($reserve->getPaymentReviewStatus() == PENDING_APPROVAL){
+            $reserve->setKeeperReviewStatus(REJECTED);
+            $this->reserveDAO->Update($reserve);
+        }else{
+            echo "<script> if(confirm('Cannot change status once payed!')); </script>";
+        }
         header("location:" . FRONT_ROOT . "Reserve/RetrievePendingReserves");
     }
 
@@ -81,28 +86,41 @@ class ReserveController
 
     public function DeleteReserve($reserveId)
     {
-        $this->reserveDAO->Delete($reserveId);
-        header("location:" . FRONT_ROOT . "Reserve/RetrievePendingReserves");
+        $event = $this->reserveDAO->GetReserve($reserveId);
+
+        if($event->getPaymentReviewStatus() == APPROVED){
+            echo "<script> if(confirm('Cannot delete once payed!')); </script>";
+            header("location:" . FRONT_ROOT . "Reserve/RetrievePendingReserves");
+        }else{
+            $this->reserveDAO->Delete($reserveId);
+            header("location:" . FRONT_ROOT . "Reserve/RetrievePendingReserves");
+        }
     }
 
     public function PaidOut($reserveId){
         $event = $this->reserveDAO->GetReserve($reserveId);
-        $keeper = $this->keeperDAO->GetKeeper($event->getKeeper());
-        $petToKeep = $this->petDAO->GetPet($event->getPets());
 
-        $keeperDateStart = new DateTime($event->getStartDate());
-        $keeperDateFinal = new DateTime($event->getFinalDate());
-        $keeperInterval = $keeperDateStart->diff($keeperDateFinal);
+        if($event->getPaymentReviewStatus() == APPROVED){
+            echo "<script> if(confirm('Already payed!')); </script>";
+            header("location:" . FRONT_ROOT . "Reserve/RetrievePendingReserves");
+        }else{
+            $keeper = $this->keeperDAO->GetKeeper($event->getKeeper());
+            $petToKeep = $this->petDAO->GetPet($event->getPets());
 
-        require_once(VIEWS_PATH . "payment-view.php");
+            $keeperDateStart = new DateTime($event->getStartDate());
+            $keeperDateFinal = new DateTime($event->getFinalDate());
+            $keeperInterval = $keeperDateStart->diff($keeperDateFinal);
 
+            require_once(VIEWS_PATH . "payment-view.php");
+        }
     }
 
     public function FinishPayment($reserveId){
         $reserve = $this->reserveDAO->GetReserve($reserveId);
         $reserve->setPaymentReviewStatus(APPROVED);
-        $this->keeperDAO->RemoveFreeTimePeriod($reserve->getKeeper(), $reserve->getStartDate(), $reserve->getFinalDate());
+        $keeperObject = $this->keeperDAO->GetKeeper($reserve->getKeeper());
+        $this->keeperDAO->RemoveFreeTimePeriod($keeperObject, $reserve->getStartDate(), $reserve->getFinalDate());
         $this->reserveDAO->Update($reserve);
-        $this->RetrieveMadedReserves();
+        header("location:" . FRONT_ROOT . "Reserve/RetrievePendingReserves");
     }
 }
